@@ -7,11 +7,14 @@
   "license": "ISC",
   "repository": "git@github.com:christophehurpeau/promise-callback-factory.git",
   "homepage": "https://github.com/christophehurpeau/promise-callback-factory",
+  "engines": {
+    "node": ">=6.5.0"
+  },
   "main": "./index.js",
   "webpack:main": "./dist/webpack/index.js",
   "webpack:main-dev": "./dist/webpack-dev/index.js",
-  "webpack:node": "./lib-node6/index.js",
-  "webpack:node-dev": "./lib-node6-dev/index.js",
+  "webpack:node": "./lib-webpack-node7/index.js",
+  "webpack:node-dev": "./lib-webpack-node7-dev/index.js",
   "webpack:main-modern-browsers": "./dist/webpack-modern-browsers/index.js",
   "webpack:main-modern-browsers-dev": "./dist/webpack-modern-browsers-dev/index.js",
   "scripts": {
@@ -27,24 +30,26 @@
     "version": "pob-version",
     "watch": "pob-watch"
   },
+  "dependencies": {
+    "flow-runtime": "^0.6.1"
+  },
   "devDependencies": {
     "babel-eslint": "^7.0.0",
     "babel-plugin-add-jsdoc-annotations": "^5.1.0",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-preset-es2015-node6": "^0.4.0",
     "babel-preset-jsdoc": "^0.4.0",
-    "babel-preset-modern-browsers": "^8.0.0",
-    "eslint": "^3.12.2",
-    "eslint-config-airbnb-base": "^11.0.0",
-    "eslint-config-pob": "^11.0.0",
-    "eslint-plugin-babel": "^4.0.0",
+    "babel-preset-latest": "^6.22.0",
+    "babel-preset-latest-node": "^0.1.0",
+    "babel-preset-modern-browsers": "^8.1.1",
+    "eslint": "^3.16.1",
+    "eslint-config-airbnb-base": "^11.1.0",
+    "eslint-config-pob": "^11.1.0",
+    "eslint-plugin-babel": "^4.1.0",
     "eslint-plugin-import": "^2.2.0",
     "jsdoc": "^3.4.1",
     "komet": "^0.1.4",
-    "komet-karma": "^0.2.3",
+    "komet-karma": "^0.2.5",
     "minami": "^1.1.1",
-    "pob-babel": "^13.0.0",
-    "pob-release": "^3.0.0",
-    "tcomb-forked": "^3.4.0"
+    "pob-babel": "^16.1.0",
+    "pob-release": "^3.1.0"
   }
 }
