@@ -1,16 +1,4 @@
-'use strict';
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.resolveFromCallback = resolveFromCallback;
-exports.default = promiseCallback;
-
-var _flowRuntime = require('flow-runtime');
-
-var _flowRuntime2 = _interopRequireDefault(_flowRuntime);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+import t from 'flow-runtime';
 
 /**
  * Creates a callback that resolve or reject a promise
@@ -20,7 +8,7 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
  * @param {Function} reject reject function of the promise
  * @return {Function}
  */
-function resolveFromCallback(resolve, reject) {
+export function resolveFromCallback(resolve, reject) {
   return function (err, result) {
     if (err) {
       if (typeof err === 'string') {
@@ -48,12 +36,12 @@ function resolveFromCallback(resolve, reject) {
  * @param {Function} callback callback((done) => {})
  * @return {Promise}
  */
-function promiseCallback(callback) {
-  let _callbackType = _flowRuntime2.default.function();
+export default function promiseCallback(callback) {
+  let _callbackType = t.function();
 
-  const _returnType = _flowRuntime2.default.return(_flowRuntime2.default.ref('Promise'));
+  const _returnType = t.return(t.ref('Promise'));
 
-  _flowRuntime2.default.param('callback', _callbackType).assert(callback);
+  t.param('callback', _callbackType).assert(callback);
 
   return _returnType.assert(new Promise((resolve, reject) => {
     callback(resolveFromCallback(resolve, reject));
